// src/payload-types.ts
export interface StatsBlock {
  items?:
    | {
        /**
         * 例如: 20+, 500, 1000+
         */
        value: string;
        /**
         * 例如: 年, 项, 家
         */
        unit?: string | null;
        label: string;
        icon?: ('Calendar' | 'Users' | 'Award' | 'TrendingUp' | 'Factory' | 'Globe') | null;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'stats';
}